feat: require user authentication
To streamline the code and ease implementation, the `connection` class
has been refactored slightly:

1. `connection::respond` has been made a function template, taking any
   type as an argument for the `message` parameter. This allows passing
   any Protobuf type that can be serialized to an array, instead of
   strictly requiring a `ResponseContainer`.
2. For the overloaded `connection::respond` method that took a
   `binary_data_view` the parameter type has been changed to be `const
   binary_data &`. This was necessary so that no implicit conversion
   from `binary_data` to `binary_data_view` is performed when using the
   return value of `database_wrapper::get_response_data_raw`.
   A function overload is only chosen over a function template if the
   parameter type is an exact match (i.e., no implicit conversion is
   required). So, in order for the compiler to pick the function
   overload, the implicit conversion needed to be prevented.

Since the `user` struct now has `binary_data` attributes,
`database_wrapper.hpp` had to be included. This required forward
declarations in order to avoid a dependency loop. Also, the `user.hpp`
header had to be included in multiple files because of this change.

A default user is created with the database initialization. The user
name is "Standard User" and the password is "pass". It can be used to
run examples that require client-server communication, for example.

Additionally, for the time being, a shell script to add users to the
database is included with this commit. It can be used like so:
```bash
create-users.sh "User Name" "password"
```
Note that the script has to been run as a Postgres user that can write
to the database, e.g. like so:
```bash
sudo -su postgres create-user.sh "User Name" "password"
```

Furthermore, the script requires `xxd` (commonly bundled with the Vim
text editor) to be installed.